% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrackll.R
\docType{methods}
\name{createTrackll}
\alias{createTrackll}
\title{createTrackll}
\usage{
createTrackll(folder, input = 0, interact = F, merge = F, ab.track = F, mask = F, cores = parallel::detectCores(logical=F), censorSingle = F, frameRecord = T)
}
\arguments{
\item{folder}{Full path output file folder (if they are .txt, ensure that they are either all Diatrack or all SlimFast)}

\item{input}{Input file type (Diatrack .txt file = 1; Diatrack .mat session file = 2; ImageJ .csv file = 3; SlimFast .txt file = 4)}

\item{merge}{An logical indicate if the output list should be merged into one. Default merge = FALSE, output list is divided by file names.}

\item{ab.track}{Use absolute coordinates for tracks}

\item{mask}{A logical indicate if image mask should be applied to screen tracks. Default False. Note the mask file should have the same name as the Diatrack output txt file with a "_MASK.tif" ending. Users can use plotMask() and plotTrackOverlay() to see the mask and its effect on screening tracks.}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralelled.}

\item{censorSingle}{Remove and censor trajectories that do not have a recorded next/previous frame (trajectories that appear for only one frame)}

\item{frameRecord}{add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames)}
}
\description{
take in Diatrack (.txt or .mat), ImageJ (.csv), or SlimFast (.txt) input from a folder to output a list of track lists with the option to merge, mask, censor, record frames, and use multiple cores.
}
\details{
NOTE: Diatrack .txt and ImageJ (.csv) input does not have the ability create a frame record and censor/uncensor, it will only extract whatever has been done by default to the .txt file.

 The naming scheme for the list of track list is as follows:
 
 Track List: [full name of input file]
 Track: 

[Last five characters of the file name].[Start frame #].[Length].[Track #].[Index # (will differ from Track # when merging)]

(Note: The last five characters of the file name, excluding the extension, cannot contain “.”)
}
\examples{

}
