% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTrackll.R
\docType{methods}
\name{exportTrackll}
\alias{exportTrackll}
\alias{.exportRowWise}
\title{exportTrackll}
\usage{
exportTrackll(trackll, cores = 1)

.exportRowWise(track.list)
}
\arguments{
\item{track.list}{a single track list}

\item{trackll}{a list of track lists}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralelled.}
}
\description{
take in a list of track lists (trackll) and export it into row-wise (ImageJ/MOSAIC) .csv files in the working directory
}
\details{
##' The naming scheme for each export uses the ISO 8601 format, as follows:

[yy-MM-ddTHH-mm-ss][Last five characters of the file name].csv

The reason why ImageJ/MOSAIC style .csv export was chosen is because it fully preserves all information, while maintaining relatively short computation time and the ability to easily read it in Excel or etc.

In order to import this .csv export back into a trackll at any point (while preserving all information), select option input 3 in createTrackll.

If the track list does not have a fourth frame record column (not recommended), it will just output the start frame of each track instead

It is not recommended that exportTrackll be run on merged list of track lists (trackll).
Also, ensure that the input trackll is a list of track lists and not just a trackl track list
}
