% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkSkippedFrames.R
\docType{methods}
\name{linkSkippedFrames}
\alias{linkSkippedFrames}
\title{linkSkippedFrames}
\usage{
linkSkippedFrames(track.list, tolerance, maxSkip)
}
\arguments{
\item{track.list}{A track list (a list of trajectory data frames)}

\item{tolerance}{distance tolerance level measured in pixels after the frame skip}

\item{maxSkip}{maximum number of frames a trajectory can skip}
}
\description{
link trajectories skipped (or do not appear for) a number of frames
}
\details{
Given user input for a tolerance level to limit how far the next point after the skip can deviate from the last point in pixel distance 
and a maximum number of frame skips possible, all trajectories falling within these parameters are automatically linked, renamed, and ordered accordingly. 

The naming scheme for each linked track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #].[# of links]

(Note: The last five characters of the file name, excluding the extension, cannot contain “.”)

Although not required, in order for the output to have a frame record column (recommended), the input must have one as well.

For a maxSkip example, if the maxSkip for a trajectory ending in frame 7 was 3, the next linked trajectory can start up to a maximum frame of 11)
}
\examples{
#Basic function call of linkSkippedFrames
trackll.linked <- linkSkippedFrames(trackll, tolerance = 4.00, maxSkip = 5)

#Option to output .csv files after processing the linked track lists
outputRowWise(trackll.linked)
outputColWise(trackll.linked)
}
