% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDiaSessions.R
\docType{methods}
\name{readDiaSessions}
\alias{readDiaSessions}
\alias{.readDiaSessions}
\title{readDiaSessions}
\usage{
.readDiaSessions(file, interact = T, ab.track = F, frameRecord = T)

removeFrameRecord(track.list)
}
\arguments{
\item{file}{Full path to Diatrack .mat session file}

\item{interact}{Open menu to interactively choose file}

\item{ab.track}{Use absolute coordinates for tracks}

\item{frameRecord}{add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames)}
}
\description{
take in a Diatrack .mat session file as input, along with several other user-configurable parameters and output options, to return a track list of all the trajectories found in the session file
}
\details{
The naming scheme for each track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #]

(Note: The last five characters of the file name, excluding the extension, cannot contain “.”)

removeFrameRecord is a helper script aims to make track lists with a fourth frame record column backwards compatible 
with other smt functions that rely on track lists with only three columns for the xyz-coordinates. 
The fourth column is simply removed from the given track list.
}
\examples{
#Basic function call of .readDiaSessions
trackl <- .readDiaSessions(interact = T)

#Option to output .csv files after processing the track lists
.exportRowWise(trackl)

}
