% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSlimFast.R
\docType{methods}
\name{readSlimFast}
\alias{readSlimFast}
\alias{.readSlimFast}
\title{readSlimFast}
\usage{
.readSlimFast(file, interact = F,  ab.track = F, censorSingle = F, frameRecord = T)
}
\arguments{
\item{file}{Full path to SlimFast .txt session file}

\item{interact}{Open menu to interactively choose file}

\item{ab.track}{Use absolute coordinates for tracks}

\item{censorSingle}{Remove and censor trajectories that appear for only one frame}

\item{frameRecord}{add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames)}
}
\description{
take in a SlimFast .txt session file as input, along with several other user-configurable parameters and output options, to return a track list of all the trajectories
}
\details{
The naming scheme for each track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #]

(Note: The last five characters of the file name, excluding the extension, cannot contain “.”)
}
\examples{
#Basic function call of .readSlimFast
trackl <- .readSlimFast(interact = T)

#Function call of .readSlimFast with censoring without a frame record and output to .csv files
trackl2 <- .readSlimFast(interact = T, censorSingle = T, frameRecord = F)

#Option to output .csv files after processing the track lists
.exportRowWise(trackl)
.exportColWise(trackl)

#To find your current working directory
getwd()

#Remove default fourth frame record column
trackll.removed <- removeFrameRecord(trackl)

}
