% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTrackll.R
\docType{methods}
\name{exportTrackll}
\alias{exportTrackll}
\alias{.exportRowWise}
\title{exportTrackll}
\usage{
exportTrackll(trackll, rowWise = T, colWise = T, cores = 1)

.exportRowWise(track.list)

.exportColWise(track.list)
}
\arguments{
\item{track.list}{a single track list}

\item{trackll}{a list of track lists}

\item{rowWise}{option to use Image-J style row-wise output in .csv files}

\item{colWise}{option to use Diatrack style col-wise output in .csv files}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralelled.}
}
\description{
take in a list of track lists (trackll) and export it into row-wise and/or column-wise .csv files in the working directory
}
\details{
For .exportRowWise, if the track list does not have a fourth frame record column, it will just output the start frame of each track instead

It is not recommended that exportTrackll be run on merged list of track lists (trackll).

Ensure that the input trackll is a list of track lists and not just a trackl track list
}
