% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTrackll.R
\docType{methods}
\name{exportTrackll}
\alias{exportTrackll}
\alias{.exportRowWise}
\title{exportTrackll}
\usage{
.exportRowWise(track.list)
}
\details{
outputRowWise is only compatible with track lists with the fourth frame record column.

outputColWise is compatible with track lists both with or without the fourth frame record column.
}
